Update to the layout, colours and the label sizes and a bug fix to the redshifts
I have updated the layout of the dendogram and the colours of the lines and
blocks showing if it is a merging or interacting branch. In addition size of the
text labels have been increased to make the plot clearer all as per referee's comments.
There is also a bug fix to the redshift in the conversion of the VELOCIraptor and
AHF format as it was set to be a interger, but is now set to be a float.