Updated size scaling and user definable options and enabled highlighting of WhereWolf halos
Updated how the sub-panels scale with the main panel so the sizes of a halo is the same on
all panels. Also made the fontsize of the maxsize data and the units of Eucleadian distance
user defineable. Enabled the convToETF to produce an extra dataset as a flag if a halo is
a WhereWolf halo or not and then it can be shown in the dendogram. This is switched on by
setting WWflag = 1 in both convToEFT.cfg and plot_config.cfg.