Removed the need for any units and corrected the unit conversion for VELOCIraptor
ETF will have units 10^10 Msun for masses, Mpc for distances and km/s for velocities
I have updated the conversion tool for VELOCIraptor to always do this. I have also
added the units to the header so the file is self contained and attributes for each
dataset so it knows its orginal name in the catalogue.