Added a another option and a further fix to the warning shown and a bug fix
Added a option so the branch label can be shown or not and updated the
plotting routine so that if there is only 1 depth shown a colorbar
is not used as this gave out a warning before. A bug fix to the plotting
routine where not all the labels were shown for the last panel.