Enabled the scaling of sub-branches and bug fix to conversion tool
Since sub-branches can be small relative to the branch of interest, I have
added a subBranchSizeFactor which scales the sub-branches. This factor is
shown in the bottom right of the dendogram. In addition there is a bug fix
to the conversion tool for the AHF and Rockstar catalogue field names.