Split up the size label and added extra config option
I have split up the size label into the dataset name and units as
this makes it easier to put the max size label on the dendogram.
In addition I have added an option so a user can change the amount
of spaces between the max size label in the top left and the main
pannels max size. This should just be a temporary fix while a
better option is being found that scales nicely with lots of
branches.